requirements.txt
Flask==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
Werkzeug==2.0.1
click==8.0.1
itsdangerous==2.0.1
packaging==21.0
pip==19.0.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
setuptools==40.8.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
gunicorn==20.1.0